requirements.txt
azure-common
azure-cosmosdb-nspkg
azure-cosmosdb-table
azure-nspkg
Flask
Flask-Cors
